requirements.txt
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
playwright>=1.40.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
colorama>=0.4.6
tenacity>=8.2.0
openai>=1.0.0
